requirements.txt
pandas
boto3
sqlalchemy
pymysql
pyarrow
python-dotenv